+++ src/main/java/com/pedrolucas/Agaply/dto/itemvenda/ItemVendaResponseDTO.java
@@ -5,7 +5,7 @@ import java.math.BigDecimal;
 public record ItemVendaResponseDTO(
         Long id,
         String produtoNome,
-        BigDecimal quantidade,
+        int quantidade,
         BigDecimal precoUnitario,
         BigDecimal subtotal
 ) {
